pyproject.toml
[tool.poetry]
name = "ccb"
version = "0.1.0"
description = "Random pairing bot for virtual coffee"
authors = ["Abhinav Tushar <abhinav@lepisma.xyz>"]

[tool.poetry.dependencies]
python = "^3.8"
docopt = "^0.6.2"
slackclient = "^2.8.2"
pydash = "^4.8.0"

[tool.poetry.dev-dependencies]
pytest = "^6.0.2"

[tool.poetry.scripts]
ccb = "ccb.cli:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
